feat: display calendar at home
