+++ frontend/package.json
@@ -13,7 +13,7 @@
     "@reduxjs/toolkit": "^1.7.1",
     "axios": "^0.24.0",
     "clsx": "latest",
-    "dayjs": "^1.10.7",
+    "date-fns": "^2.28.0",
     "formik": "^2.2.9",
     "googleapis": "^92.0.0",
     "jsonwebtoken": "^8.5.1",
@@ -39,6 +39,7 @@
     "@types/jsonwebtoken": "^8.5.6",
     "@types/node": "16.11.11",
     "@types/react": "17.0.37",
+    "@types/react-big-calendar": "^0.36.2",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "4.0.0",
     "autoprefixer": "^10.4.0",
