+++ frontend/src/component/UserCalendar/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserCalendar";
