+++ frontend/src/component/WelcomePanel/index.ts
@@ -1,0 +1,1 @@
+export * from "./WelcomePanel";
