+++ frontend/src/component/index.tsx
@@ -2,3 +2,5 @@ export * from "./Copyright";
 export * from "./MeetupForm";
 export * from "./Link";
 export * from "./ProTip";
+export * from "./UserCalendar";
+export * from "./WelcomePanel";
