+++ frontend/src/services/backend/index.ts
@@ -1,0 +1,2 @@
+export * from "./backend";
+export * from "./model";
